Refactor tests; import instead of use TestHelpers
