+++ test/dreadnought/linear_algebra/csys_test.exs
@@ -5,7 +5,7 @@ defmodule LinearAlgebra.CsysTest do
   use ExUnit.Case, async: true
   use Dreadnought.PositionOrientationSize
   use Dreadnought.LinearAlgebra
-  use Dreadnought.TestHelpers
+  import Dreadnought.TestHelpers
 
   describe "vector coordinate system" do
